auto

+++ src/main.py
@@ -39,7 +39,7 @@ parser.add_argument('-e', '--epochs', type=int, default=10,
                     help='Epochs')
 parser.add_argument('-b', '--batch_size', type=int, default=8,
                     help='Batch Size')
-parser.add_argument('-l', '--learning_rate', type=int, default=0.0001,
+parser.add_argument('-l', '--learning_rate', type=float, default=0.0001,
                     help='Learning Rate')
 parser.add_argument('-w', '--weighted_loss', action='store_true',
                     help='Weighted Loss')
